+++ 2016/18.cpp
@@ -2,8 +2,8 @@
 #include <chrono>
 #include <iostream>
 
-int count_traps(const std::string& input, int nrows) {
-  int trapcount = 0;
+size_t count_traps(const std::string& input, size_t nrows) {
+  size_t trapcount = 0;
 
   // store row in bitset that's larger than row width
   // so we don't have to do any bounds checking inside the loop
@@ -13,7 +13,7 @@ int count_traps(const std::string& input, int nrows) {
   prev_row[0] = false;
   prev_row[101] = false;
 
-  for (int c = 0; c < 100; c++) {
+  for (size_t c = 0; c < 100; c++) {
     prev_row[c + 1] = input[c] == '^' ? true : false;
     if (input[c] == '^') {
       trapcount += 1;
@@ -21,14 +21,14 @@ int count_traps(const std::string& input, int nrows) {
   }
 
   int tiles;
-  for (int r = 1; r < nrows; r++) {
-    for (int c = 1; c <= 100; c++) {
+  for (size_t r = 1; r < nrows; r++) {
+    for (size_t c = 1; c <= 100; c++) {
       tiles = (prev_row[c - 1] << 2) + (prev_row[c] << 1) + prev_row[c + 1];
       cur_row[c] = (tiles == 0b110 || tiles == 0b011 || tiles == 0b100 ||
                     tiles == 0b001);
     }
 
-    trapcount += static_cast<int>(cur_row.count());
+    trapcount += cur_row.count();
     prev_row = cur_row;
   }
 
@@ -37,8 +37,8 @@ int count_traps(const std::string& input, int nrows) {
 
 int main() {
   auto tstart = std::chrono::high_resolution_clock::now();
-  int pt1 = 0;
-  int pt2 = 0;
+  size_t pt1 = 0;
+  size_t pt2 = 0;
 
   std::string input;
   std::getline(std::cin, input);
